refactor(identity): split the /v1 admission middleware into middleware.go
handlers.go crossed the 500-LOC cap after the connector-callback exemption;
move publicPaths + isConnectorOAuthCallback + Middleware into middleware.go
(one concept per file). No behaviour change.
